Fixed Losing Robot Code / Immediate Disable
The "wheels" Spark in the Lift subsystem was never .reset() in Lift.cpp, meaning it was a null pointer. The LiftDefault command tried to set "wheels" to 0.0, but since it was a null pointer it caused an immediate code crash upon enabling.

NOTE: Lift->wheels currently won't work because it is set to use PWM 22. PWM ports need refactoring to make space for the "wheels."